Modified to execute under Python3 so that it can be used with the newer WeeWx (tested with v4.4.0).  The utility 2to3 was used initially but some problems remained that prevented the generator from working as intended.
Credit to author of ADDS_METAR as I borrowed a snippet of code from their work to perform the retrieval of the METAR html.  I must apologize for forgetting the authors name - my bad.  This code block was setup to request
the html upto 3 times before giving up and reporting an error.

I (perhaps incorrectly) assume that the str.find("Output produced by METARs form") is done to verify the format of the returned html body.  The string wasn't found in line[0] so I took the liberty of doing myfile.find()
in its place.  That worked as expected.

Another issue that wasn't so easy to discover was that the variable "destination_dir" was an empty string.  That coupled with 'with open() as f' had me scratching my head for some days.  It seems that the backup file
could not be opened in the default directory (whatever that is) for write or for read and the generator was quick to exit without reporting any errors other than it failed to open the backup file.  To correct the
issue, and I expect Dani may find a better way to fix it, I simply substituted the string "/tmp/" for the variable destination_dir during filename construction.  The backup file(s) get created in the /tmp directory
and are available as expected when a request to aviationweather.gov fails EXCEPT following a system reboot.  This is because the system will clear /tmp during the boot sequence removing any backup files that might
exist there.

A few additional LOG_INFO statements were also added.

Thanks to Dani for the original program.  I certainly think it is a useful addition to my weather station particularly when my grandchildren are learning there is more to a weather report than what one gets from a talking
head on the television.

Enjoy!